requirements.txt
langdetect==1.0.9
pdfplumber==0.11.0
python-docx==1.1.0
beautifulsoup4==4.12.3
sentence-transformers==2.3.1
chromadb==0.4.22
ollama==0.1.7
python-magic==0.4.27
thefuzz==0.22.1
python-Levenshtein==0.23.0
lxml==5.1.0
reportlab==4.0.7
matplotlib==3.8.2

